subscriber/requirements.txt
psycopg2-binary==2.9.9
paho-mqtt==2.0.0
python-dotenv==1.0.1